Update main.py
Made the GUI better and made the Python file have some sort of redundancy. Instead of just killing itself when a module is missing and making the user install it, it automagically installs the required packages. If it can't then it will send the error to the user telling them to contact the files creator.